+++ compiler/src/x86_arch_full.ml
@@ -120,6 +120,7 @@ module X86_core = struct
     | STC -> false (* Not DOIT *)
     | SUB _ -> true
     | TEST _ -> true
+    | TZCNT _ -> false (* Not DOIT *)
     | VAESDEC _ -> true
     | VAESDECLAST _ -> true
     | VAESENC _ -> true